+++ test/v2.amazonecr.test.js
@@ -84,7 +84,7 @@ test('v2 amazonecr', function (tt) {
         noauthClient.listTags(function (err) {
             t.ok(err);
             t.equal(err.statusCode, 401, 'Expect a 401 status code');
-            t.equal(String(err.message).trim(), 'Not Authorizied');
+            t.equal(String(err.message).trim(), 'Not Authorized');
             t.end();
         });
     });
@@ -94,6 +94,7 @@ test('v2 amazonecr', function (tt) {
      */
     tt.test('  createClient', function (t) {
         client = drc.createClientV2({
+            maxSchemaVersion: 2,
             name: CONFIG.repo,
             username: CONFIG.username,
             password: CONFIG.password,
@@ -123,18 +124,22 @@ test('v2 amazonecr', function (tt) {
     });
 
     /*
-     *  {
-     *      "name": <name>,
-     *      "tag": <tag>,
-     *      "fsLayers": [
-     *         {
-     *            "blobSum": <tarsum>
-     *         },
-     *         ...
-     *      ],
-     *      "history": <v1 images>,
-     *      "signature": <JWS>
-     *  }
+     * {
+     *   "schemaVersion": 2,
+     *   "mediaType": "application/vnd.docker.distribution.manifest.v2+json",
+     *   "config": {
+     *     "mediaType": "application/octet-stream",
+     *     "size": 1459,
+     *     "digest": "sha256:2b8fd9751c4c0f5dd266fc...01"
+     *   },
+     *   "layers": [
+     *     {
+     *       "mediaType": "application/vnd.docker.image.rootfs.diff.tar.gzip",
+     *       "size": 667590,
+     *       "digest": "sha256:8ddc19f16526912237dd8af...a9"
+     *     }
+     *   ]
+     * }
      */
     var manifest;
     var manifestDigest;
@@ -147,13 +152,12 @@ test('v2 amazonecr', function (tt) {
             manifestDigest = res.headers['docker-content-digest'];
             t.equal(manifestDigest, undefined, 'no docker-content-digest');
             t.ok(manifest);
-            t.equal(manifest.schemaVersion, 1);
-            t.equal(manifest.name, repo.remoteName);
-            t.equal(manifest.tag, CONFIG.tag);
-            t.ok(manifest.architecture);
-            t.ok(manifest.fsLayers);
-            t.ok(manifest.history[0].v1Compatibility);
-            t.ok(manifest.signatures[0].signature);
+            t.equal(manifest.schemaVersion, 2);
+            t.ok(manifest.config);
+            t.ok(manifest.config.digest, manifest.config.digest);
+            t.ok(manifest.layers);
+            t.ok(manifest.layers.length > 0);
+            t.ok(manifest.layers[0].digest);
             t.end();
         });
     });
@@ -168,7 +172,7 @@ test('v2 amazonecr', function (tt) {
     });
 
     tt.test('  headBlob', function (t) {
-        var digest = manifest.fsLayers[0].blobSum;
+        var digest = manifest.layers[0].digest;
         client.headBlob({digest: digest}, function (err, ress) {
             t.ifErr(err);
             t.ok(ress);
@@ -196,10 +200,9 @@ test('v2 amazonecr', function (tt) {
             t.equal(last.statusCode, 200);
 
             // Content-Type:
-            // - docker.io gives 'application/octet-stream', but amazon isn't so
-            //   nice for a HEAD request, it just returns text/plain.
+            // - Note that docker.io gives 'application/octet-stream'
             t.equal(last.headers['content-type'],
-                'text/plain; charset=utf-8');
+                'application/vnd.docker.image.rootfs.diff.tar.gzip');
 
             t.ok(last.headers['content-length']);
             t.end();
@@ -230,7 +233,7 @@ test('v2 amazonecr', function (tt) {
     });
 
     tt.test('  createBlobReadStream', function (t) {
-        var digest = manifest.fsLayers[0].blobSum;
+        var digest = manifest.layers[0].digest;
         client.createBlobReadStream({digest: digest},
                 function (err, stream, ress) {
             t.ifErr(err);