+++ test/v2.jfrogartifactory.test.js
@@ -19,8 +19,7 @@
  *              "repo": "trentm.artifactoryonline.com/busybox",
  *              "username": "admin",
  *              "password": "(your password)",
- *              "tag": "latest",
- *              "searchTerm": "busy"
+ *              "tag": "latest"
  *          }
  *      }
  *
@@ -68,6 +67,7 @@ test('v2 jfrog artifactory private repo (' + CONFIG.repo + ')', function (tt) {
 
     tt.test('  createClient', function (t) {
         client = drc.createClientV2({
+            maxSchemaVersion: 2,
             name: CONFIG.repo,
             username: CONFIG.username,
             password: CONFIG.password,
@@ -88,12 +88,11 @@ test('v2 jfrog artifactory private repo (' + CONFIG.repo + ')', function (tt) {
 
     tt.test('  ping', function (t) {
         client.ping(function (err, body, res) {
-            // With JFrog Artifactory, we actually get a successful ping,
-            // because (I guess) only basic auth is required here.
-            t.ifErr(err);
+            // Expect a 401 (Not Authorized) error here.
+            t.ok(err);
             t.ok(res, 'have a response');
             if (res) {
-                t.equal(res.statusCode, 200);
+                t.equal(res.statusCode, 401);
                 t.equal(res.headers['docker-distribution-api-version'],
                     'registry/2.0');
             }
@@ -159,18 +158,22 @@ test('v2 jfrog artifactory private repo (' + CONFIG.repo + ')', function (tt) {
     });
 
     /*
-     *  {
-     *      "name": <name>,
-     *      "tag": <tag>,
-     *      "fsLayers": [
-     *         {
-     *            "blobSum": <tarsum>
-     *         },
-     *         ...
-     *      ],
-     *      "history": <v1 images>,
-     *      "signature": <JWS>
-     *  }
+     * {
+     *   "schemaVersion": 2,
+     *   "mediaType": "application/vnd.docker.distribution.manifest.v2+json",
+     *   "config": {
+     *     "mediaType": "application/octet-stream",
+     *     "size": 1459,
+     *     "digest": "sha256:2b8fd9751c4c0f5dd266fc...01"
+     *   },
+     *   "layers": [
+     *     {
+     *       "mediaType": "application/vnd.docker.image.rootfs.diff.tar.gzip",
+     *       "size": 667590,
+     *       "digest": "sha256:8ddc19f16526912237dd8af...a9"
+     *     }
+     *   ]
+     * }
      */
     var manifest;
     var manifestDigest;
@@ -180,13 +183,12 @@ test('v2 jfrog artifactory private repo (' + CONFIG.repo + ')', function (tt) {
             manifest = manifest_;
             manifestDigest = res.headers['docker-content-digest'];
             t.ok(manifest);
-            t.equal(manifest.schemaVersion, 1);
-            t.equal(manifest.name, repo.remoteName);
-            t.equal(manifest.tag, CONFIG.tag);
-            t.ok(manifest.architecture);
-            t.ok(manifest.fsLayers);
-            t.ok(manifest.history[0].v1Compatibility);
-            t.ok(manifest.signatures[0].signature);
+            t.equal(manifest.schemaVersion, 2);
+            t.ok(manifest.config);
+            t.ok(manifest.config.digest, manifest.config.digest);
+            t.ok(manifest.layers);
+            t.ok(manifest.layers.length > 0);
+            t.ok(manifest.layers[0].digest);
             t.end();
         });
     });
@@ -215,7 +217,7 @@ test('v2 jfrog artifactory private repo (' + CONFIG.repo + ')', function (tt) {
     });
 
     tt.test('  headBlob', function (t) {
-        var digest = manifest.fsLayers[0].blobSum;
+        var digest = manifest.layers[0].digest;
         client.headBlob({digest: digest}, function (err, ress) {
             t.ifErr(err);
             t.ok(ress);
@@ -261,7 +263,7 @@ test('v2 jfrog artifactory private repo (' + CONFIG.repo + ')', function (tt) {
     });
 
     tt.test('  createBlobReadStream', function (t) {
-        var digest = manifest.fsLayers[0].blobSum;
+        var digest = manifest.layers[0].digest;
         client.createBlobReadStream({digest: digest},
                 function (err, stream, ress) {
             t.ifErr(err);